restructrue app files.
